// projects/codx-ac/src/lib/settings/fiscal-periods/fiscal-periods-add/model/FiscalPeriodsAutoCreate.model.ts
export class FiscalPeriodsAutoCreate{
    periodName: any;
    periodName2: any;
    periodType: any;
    fiscalYear: any;
    startDate: any;
    endDate: any;
    periodControl: any;
    moduleControl: any;
    createdOn: any;
    createdBy: any;
    modifiedOn: any;
    modifiedBy: any;
    oldFiscalYear: any;
    fiscalPeriodControl: any;
    period: any;
    periodUoM: any;
}